+++ src/Layout/Helper/EdgeProxy.php
@@ -23,40 +23,13 @@
  * SOFTWARE.
  */
 
-namespace PhpDA\Command;
+namespace PhpDA\Layout\Helper;
 
-use PhpDA\Command\MessageInterface as Message;
-use PhpDA\Plugin\FactoryInterface;
-use PhpDA\Plugin\Loader;
-use Symfony\Component\Console\Application;
-use Symfony\Component\Yaml\Parser;
+use Fhaculty\Graph\Edge\Directed;
 
-class ApplicationFactory implements FactoryInterface
+/**
+ * @property \PhpDA\Entity\Location[] $locations
+ */
+class EdgeProxy extends Directed
 {
-    /**
-     * @return Application
-     */
-    public function create()
-    {
-        $app = new Application(Message::NAME, Message::VERSION);
-        $app->setDefaultCommand(Message::COMMAND);
-        $app->add($this->createAnalyzeCommand());
-
-        return $app;
-    }
-
-    /**
-     * @return Analyze
-     */
-    protected function createAnalyzeCommand()
-    {
-        $command = new Analyze(Message::COMMAND);
-
-        $command->setHelp(Message::HELP);
-        $command->setDescription(Message::NAME . ' (' . Message::VERSION . ')');
-        $command->setConfigParser(new Parser);
-        $command->setStrategyLoader(new Loader);
-
-        return $command;
-    }
 }