+++ src/Writer/Strategy/Json.php
@@ -25,7 +25,7 @@
 
 namespace PhpDA\Writer\Strategy;
 
-use Fhaculty\Graph\Graph;
+use PhpDA\Layout\Graph;
 use PhpDA\Layout\GraphViz;
 
 /**
@@ -61,7 +61,6 @@ class Json implements StrategyInterface
 
         $vertices = $graph->getVertices();
         foreach ($vertices as $from) {
-            /** @var \Fhaculty\Graph\Vertex $from */
             $toVertices = $from->getVerticesEdgeTo();
             foreach ($toVertices->getVerticesDistinct() as $to) {
                 /** @var \Fhaculty\Graph\Vertex $to */